requirements.txt
streamlit==1.52.2

# PyTorch (CPU only)
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1

numpy==2.2.6
Pillow==12.0.0
PyYAML==6.0.3
scipy==1.15.3

opencv-python-headless==4.12.0.88

onnx==1.20.0
onnxruntime==1.23.2
onnx2pytorch==0.5.3

gdown
